+++ src/main/java/me/emafire003/dev/lightwithin/items/LightItems.java
@@ -8,7 +8,6 @@ import net.minecraft.item.*;
 import net.minecraft.registry.Registries;
 import net.minecraft.registry.Registry;
 import net.minecraft.registry.RegistryKey;
-import net.minecraft.util.Identifier;
 import net.minecraft.util.Rarity;
 
 import java.util.UUID;